third_party/blink/web_tests/platform/generic/fast/css/id-or-class-before-stylesheet-expected.txt

Test that elements before inline stylesheets get their style updated properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(document.getElementById("testid")).backgroundColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementsByClassName("testclass")[0]).backgroundColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementById("testID2")).backgroundColor is "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.getElementsByClassName("testClass2")[0]).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

id test
class test
case insensitive id test
case insensitive class test
